312551011 Lab4

+++ lab4/kernel/bsp_port/uart.h
@@ -5,7 +5,7 @@ void uart_init();
 void uart_send(unsigned int c);
 char uart_getc();
 void uart_puts(const char *s);
-void uart_hex(unsigned int d);
+void uart_hex(unsigned long long d);
 
 void uart_enable_tx_interrupt();
 void uart_disable_tx_interrupt();
